refactored component tree, and created render props to pass cohort data to the cohort view, now pieces are all connected
