+++ src/components/CohortSelector.jsx
@@ -1,27 +1,8 @@
 import React, { Component } from 'react';
-import axios from 'axios'
 
-class CohortSelector extends Component {
-    state = { 
-        cohorts: [],
-        selectedCohort: '',
-        location: '' 
-    }
-
-    componentDidMount(){
-        axios.get('/api/getAllCohorts').then(res => {
-            this.setState({cohorts: res.data})
-        })
-    }
-    updateSelectedCohort = () => (e) => {
-        this.setState({selectedCohort: e.target.value})
-    }
-    updateLocation = () => (e) => {
-        this.setState({location: e.target.value})
-    }
-    render() { 
-        const {cohorts, selectedCohort, location} = this.state
-        console.log(location)
+const CohortSelector =(props) => {
+  
+        const {cohorts, selectedCohort, location, updateLocation, updateSelectedCohort} = props
         let filteredCohorts = cohorts
         if(location){
             filteredCohorts = cohorts.slice().filter(cohort => cohort.location === location)
@@ -31,20 +12,19 @@ class CohortSelector extends Component {
         })
         return ( 
             <div>
-                <select value={location} onChange={this.updateLocation()}>
+                <select value={location} onChange={updateLocation()}>
                     <option default value=''>All Locations</option>
                     <option value='Dallas'>Dallas</option>
                     <option value='Lehi'>Lehi</option>
                     <option value='PHX'>Pheonix</option>
                     <option value='Provo'>Provo</option>
                 </select>
-                <select value={selectedCohort} onChange={this.updateSelectedCohort()}>
+                <select value={selectedCohort} onChange={updateSelectedCohort()}>
                     <option value=''>Choose a Cohort</option>
                     {cohortSelectors}
                 </select>
             </div> 
         );
     }
-}
  
 export default CohortSelector;
